requirements_drive.txt
# Bot avec Google Drive
python-telegram-bot
mistralai
python-dotenv

# Google Drive
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Lecture PDF
PyPDF2
pdfplumber

# Stockage et recherche
chromadb
langchain
langchain-community
sentence-transformers